ai-fastapi/requirements.txt
# FastAPI ve bağımlılıklar
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Kütüphaneleri
openai==1.10.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# OCR için
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80

# Veritabanı (opsiyonel)
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4
python-multipart==0.0.6  # File upload
aiohttp==3.9.1  # Async HTTP client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For async testing

# Optional - Deep Learning
# tensorflow==2.15.0
# torch==2.1.2
# transformers==4.37.0
